examples/meetings/requirements.txt
dotenv-python
duckyai